Configure Text package
