--- Text/Caesar's cipher/tests.py
@@ -1,17 +1,0 @@
-#!/usr/bin/env python3
-import unittest
-
-from caesars_cipher import decrypt, encrypt
-
-
-class Test(unittest.TestCase):
-
-    def test_encrypt(self):
-        self.assertEqual(encrypt("hello", 13), "uryyb")
-
-    def test_decrypt(self):
-        self.assertEqual(decrypt("uryyb", 13), "hello")
-
-
-if __name__ == "__main__":
-    unittest.main()
