+++ Text/Check_if_Palindrome/check_if_palindrome.py
