+++ Text/Half_the_String/half_the_string.py
