+++ Text/MD5_Hash_Generator/md5_hash_generator.py
