+++ Text/Password_Generator/password_generator.py
