+++ Text/Regex_Query_Tool/regex_query_tool.pyw
