+++ Text/Scooby_Doo/scooby_doo.py
